Refactor grading decision handling in AnswersRequiringAttentionItem and introduce TeacherGradingDecisionControls component
- Simplified the AnswersRequiringAttentionItem component by removing unused state and logic related to custom point handling.
- Introduced a new TeacherGradingDecisionControls component to encapsulate grading decision functionalities, including zero points, full points, and custom points.
- Enhanced the user experience by streamlining the grading decision submission process and improving code maintainability.
