+++ services/main-frontend/src/components/page-specific/manage/exercises/id/submissions/AnswersRequiringAttentionItem.tsx
@@ -1,25 +1,22 @@
-import { css, cx } from "@emotion/css"
+import { css } from "@emotion/css"
 import styled from "@emotion/styled"
 import { ExclamationMessage } from "@vectopus/atlas-icons-react"
 import { parseISO } from "date-fns"
-import React, { useState } from "react"
+import React, { useCallback, useState } from "react"
 import { useTranslation } from "react-i18next"
-import { usePopper } from "react-popper"
 
 import SubmissionIFrame from "../../../../submissions/id/SubmissionIFrame"
 
 import FlaggedPeerReviewAccordion from "./FlaggedPeerReviewAccordion"
-import PeerReviewAccordion from "./PeerOrSelfReviewAccordion"
+import PeerOrSelfReviewAccordion from "./PeerOrSelfReviewAccordion"
+import TeacherGradingDecisionControls from "./TeacherGradingDecisionControls"
 
 import { createTeacherGradingDecision } from "@/services/backend/teacher-grading-decisions"
 import {
   AnswerRequiringAttentionWithTasks,
   NewTeacherGradingDecision,
-  TeacherDecisionType,
 } from "@/shared-module/common/bindings"
-import Button from "@/shared-module/common/components/Button"
 import useToastMutation from "@/shared-module/common/hooks/useToastMutation"
-import ArrowDown from "@/shared-module/common/img/caret-arrow-down.svg"
 import { primaryFont } from "@/shared-module/common/styles"
 import { respondToOrLarger } from "@/shared-module/common/styles/respond"
 
@@ -37,12 +34,6 @@ const StatusPanel = styled.div`
   align-items: center;
 `
 
-const CustomPointPopup = css`
-  background-color: #e2e4e6;
-  padding: 2em;
-  z-index: 5;
-`
-
 const TopBar = styled.div`
   width: 100%;
   height: 108px;
@@ -51,29 +42,14 @@ const TopBar = styled.div`
   align-items: center;
 `
 
-const ControlPanel = styled.div`
-  background: #f5f5f5;
-  width: 100%;
-  height: 150px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-`
-
-const PLACEMENT = "bottom"
-const ARROW = "arrow"
-
 const AnswersRequiringAttentionItem: React.FC<Props> = ({
   answerRequiringAttention,
   exerciseMaxPoints,
+  refetch,
 }) => {
   const { t } = useTranslation()
   const [updatedPoints, setUpdatedPoints] = useState<number | null>(null)
-  const [open, setOpen] = useState(false)
-  const [sliderValue, setSliderValue] = useState<number>(0)
-  const [referenceElement, setReferenceElement] = useState<HTMLButtonElement | null>(null)
-  const [popperElement, setPopperElement] = useState<HTMLElement | null>(null)
-  const [arrowElement, setArrowElement] = useState<HTMLElement | null>(null)
+
   const submitMutation = useToastMutation(
     (update: NewTeacherGradingDecision) => {
       return createTeacherGradingDecision(update)
@@ -85,55 +61,16 @@ const AnswersRequiringAttentionItem: React.FC<Props> = ({
     {
       onSuccess: (data) => {
         setUpdatedPoints(data.score_given)
+        refetch()
       },
     },
   )
-
-  const { styles, attributes } = usePopper(referenceElement, popperElement, {
-    placement: PLACEMENT,
-    modifiers: [
-      { name: ARROW, options: { element: arrowElement, padding: 10 } },
-      {
-        name: "offset",
-        options: {
-          offset: [0, 20],
-        },
-      },
-    ],
-  })
-
-  const doSubmitChange = async (
-    user_exercise_state_id: string,
-    exercise_id: string,
-    action: TeacherDecisionType,
-    value?: number | undefined,
-  ) => {
-    const manual_points = value !== undefined ? value : null
-    submitMutation.mutate({
-      user_exercise_state_id,
-      exercise_id,
-      action: action,
-      manual_points: manual_points,
-      justification: null,
-      hidden: false,
-    })
-  }
-
-  const handleSubmitAndClose = (user_exercise_state_id: string, exercise_id: string) => {
-    doSubmitChange(
-      user_exercise_state_id,
-      exercise_id,
-      // eslint-disable-next-line i18next/no-literal-string
-      "CustomPoints",
-      sliderValue,
-    )
-    setOpen(false)
-  }
-
-  const handleOpenPopup = (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
-    e.preventDefault()
-    setOpen(!open)
-  }
+  const handleGradingDecisionSubmit = useCallback(
+    async (decision: NewTeacherGradingDecision) => {
+      submitMutation.mutate(decision)
+    },
+    [submitMutation],
+  )
 
   return (
     <>
@@ -183,7 +120,6 @@ const AnswersRequiringAttentionItem: React.FC<Props> = ({
                 color: white;
               `}
             >
-              {" "}
               {t("user-id")}: {answerRequiringAttention?.user_id}
             </p>
           </div>
@@ -205,9 +141,9 @@ const AnswersRequiringAttentionItem: React.FC<Props> = ({
               {updatedPoints === null ? answerRequiringAttention.score_given : updatedPoints}/
               {exerciseMaxPoints}
             </p>
-            {}
           </div>
         </TopBar>
+
         <p
           className={css`
             margin-top: 1.5em;
@@ -252,187 +188,27 @@ const AnswersRequiringAttentionItem: React.FC<Props> = ({
               </span>
             </div>
           </StatusPanel>
-          <ControlPanel>
-            <div
-              className={css`
-                margin-left: 1em;
-              `}
-            >
-              <h3
-                className={css`
-                  color: #4b4b4b;
-                  margin-bottom: 1rem;
-                `}
-              >
-                {t("grading")}
-              </h3>
-            </div>
-            <div
-              className={css`
-                display: flex;
-                align-items: center;
-              `}
-            >
-              <Button
-                className={css`
-                  font-family: ${primaryFont};
-                  font-weight: 600;
-                  font-size: 16px;
-                  margin-left: 1em;
-                  margin-right: 0.5em;
-                `}
-                size="medium"
-                variant="reject"
-                onClick={() =>
-                  doSubmitChange(
-                    answerRequiringAttention.id,
-                    answerRequiringAttention.exercise_id,
-                    // eslint-disable-next-line i18next/no-literal-string
-                    "ZeroPoints",
-                  )
-                }
-              >
-                {t("button-text-zero-points")}
-              </Button>
-              <Button
-                size="medium"
-                variant="primary"
-                className={css`
-                  margin-right: 0.5em;
-                `}
-                onClick={() =>
-                  doSubmitChange(
-                    answerRequiringAttention.id,
-                    answerRequiringAttention.exercise_id,
-                    // eslint-disable-next-line i18next/no-literal-string
-                    "FullPoints",
-                  )
-                }
-              >
-                {t("button-text-full-points")}
-              </Button>
-              <Button
-                size="medium"
-                variant="white"
-                type="button"
-                id="custom-point-button-v2"
-                ref={setReferenceElement}
-                onClick={handleOpenPopup}
-              >
-                {t("button-text-custom-points")}
-                <ArrowDown
-                  className={css`
-                    transform: scale(1.2);
-                    margin-left: 0.6em;
-                    margin-bottom: 4px;
-                  `}
-                />
-              </Button>
-            </div>
-          </ControlPanel>
+
+          <TeacherGradingDecisionControls
+            userExerciseStateId={answerRequiringAttention.id}
+            exerciseId={answerRequiringAttention.exercise_id}
+            exerciseMaxPoints={exerciseMaxPoints}
+            onGradingDecisionSubmit={handleGradingDecisionSubmit}
+          />
         </div>
       </div>
-      {open ? (
-        <div
-          id="custom-point-popup"
-          ref={setPopperElement}
-          className={cx(CustomPointPopup)}
-          // eslint-disable-next-line react/forbid-dom-props
-          style={styles.popper}
-          {...attributes.popper}
-        >
-          {/* eslint-disable-next-line react/forbid-dom-props */}
-          <div id="arrow" ref={setArrowElement} style={styles.arrow} />
-          <div
-            className={css`
-              display: flex;
-              flex-direction: row;
-              margin-bottom: 1em;
-              justify-content: space-evenly;
-            `}
-          >
-            <div
-              className={css`
-                width: 100%;
-                padding: 13px 0;
-                align-self: strech;
-              `}
-            >
-              <input
-                className={css`
-                  height: 4px;
-                `}
-                type="range"
-                min="0"
-                max={exerciseMaxPoints}
-                step={0.1}
-                value={typeof sliderValue === "number" ? sliderValue : 0.0}
-                onChange={(event) => setSliderValue(Number(event.target.value))}
-                aria-labelledby="input-slider"
-              />
-            </div>
-
-            <input
-              className={css`
-                margin-left: 1.5em;
-                max-width: 4em;
-                background: none;
-                border: 0px;
-                border-bottom: 1px solid black;
-              `}
-              value={sliderValue}
-              onChange={(event) => setSliderValue(Number(event.target.value))}
-              min="0.0"
-              step={0.1}
-              max="exerciseMaxPoints"
-              type="number"
-              aria-labelledby="input-slider"
-            />
-          </div>
-          <div>
-            <Button
-              type="button"
-              variant="white"
-              size="medium"
-              onClick={(e) => {
-                e.preventDefault()
-                setOpen(!open)
-              }}
-            >
-              {t("button-text-cancel")}
-            </Button>
-            <Button
-              size="medium"
-              variant="primary"
-              disabled={
-                sliderValue > exerciseMaxPoints ||
-                sliderValue < 0 ||
-                sliderValue.toString().length > exerciseMaxPoints.toString().length + 4
-              }
-              onClick={() =>
-                handleSubmitAndClose(
-                  answerRequiringAttention.id,
-                  answerRequiringAttention.exercise_id,
-                )
-              }
-            >
-              {t("button-text-give-custom-points")}
-            </Button>
-          </div>
-        </div>
-      ) : null}
 
       <div
         className={css`
           margin-bottom: 3rem;
         `}
       >
-        <PeerReviewAccordion
+        <PeerOrSelfReviewAccordion
           peerOrSelfReviews={answerRequiringAttention.received_peer_or_self_reviews}
           title={t("received-reviews")}
         />
 
-        <PeerReviewAccordion
+        <PeerOrSelfReviewAccordion
           peerOrSelfReviews={answerRequiringAttention.given_peer_reviews}
           title={t("given-peer-reviews-to-other-students")}
         />
